Move everything in a single directory
