+++ resources/lib/statichelper.py
