+++ resources/lib/tokenresolver.py
@@ -3,7 +3,7 @@
 # GNU General Public License v3.0 (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, unicode_literals
-from resources.lib.helperobjects import helperobjects
+from resources.lib import helperobjects
 
 try:
     from urllib.parse import urlencode, unquote
